requirements.txt
numpy>=1.26.4
matplotlib>=3.8.4
seaborn>=0.13.2
pillow>=10.1.0
